Create toc-01
Input: "aaa"

Output: String accepted! (The string starts and ends with 'a', following the DFA rules.)
Input: "abab"

Output: String rejected! (The string doesn't start with 'a' according to the DFA rules.)
Input: "aacaa"

Output: String accepted! (The string starts and ends with 'a', following the DFA rules.)
Input: "baba"

Output: String rejected! (The string doesn't start with 'a' according to the DFA rules.)
Input: "a"