Create toc-03
Input: "001"

Output: String does not belong to the CFG language! (The string does not follow the CFG rules.)
Input: "000111"

Output: String does not belong to the CFG language! (The string does not follow the CFG rules.)
Input: "1001"

Output: String does not belong to the CFG language! (The string does not start with '0' according to the CFG rules.)